In bm_timer_get_id fix error for assigning a pointer to different integer length by casting to uintptr_t first

+++ common/bm_freertos.c
@@ -161,7 +161,7 @@ BmErr bm_timer_change_period(BmTimer timer, uint32_t period_ms,
 }
 
 uint32_t bm_timer_get_id(BmTimer timer) {
-    return (uint32_t) pvTimerGetTimerID(timer);
+  return (uint32_t)(uintptr_t)pvTimerGetTimerID(timer);
 }
 
 uint32_t bm_get_tick_count(void) { return xTaskGetTickCount(); }
